nixos/systemd-boot: Rerun secrets every switch

+++ nixos/modules/system/boot/loader/systemd-boot/systemd-boot-builder.py
@@ -65,6 +65,8 @@ class CopyWriter:
     source: Path
 
     def write_boot_file(self, path: Path) -> None:
+        if path.exists():
+            return
         with tempfile.NamedTemporaryFile(
             mode="wb",
             dir=path.parent,
@@ -86,20 +88,29 @@ class InitrdWithSecretsWriter:
     initrd_secrets: Path
 
     def write_boot_file(self, path: Path) -> None:
-        with tempfile.NamedTemporaryFile(
-            mode="wb",
-            dir=path.parent,
-            delete=False,
-            prefix=path.name,
-            suffix=".tmp",
-        ) as tmp:
-            with open(self.source, mode="rb") as source_file:
-                shutil.copyfileobj(source_file, tmp)
-            tmp.flush()
-            run([self.initrd_secrets, tmp.name])
-            os.fsync(tmp.fileno())
-            tmp.close()
-            os.rename(tmp.name, path)
+        if path.exists():
+            # TODO: This is for matching old behavior. Previously, we
+            # appended secrets to initrd every time no matter what, so
+            # we must continue doing that. One potential alternative
+            # would be running the secret script on an empty file and
+            # using the hashed contents of that file as part of the
+            # key that the path name is based on.
+            run([self.initrd_secrets, path])
+        else:
+            with tempfile.NamedTemporaryFile(
+                mode="wb",
+                dir=path.parent,
+                delete=False,
+                prefix=path.name,
+                suffix=".tmp",
+            ) as tmp:
+                with open(self.source, mode="rb") as source_file:
+                    shutil.copyfileobj(source_file, tmp)
+                tmp.flush()
+                run([self.initrd_secrets, tmp.name])
+                os.fsync(tmp.fileno())
+                tmp.close()
+                os.rename(tmp.name, path)
 
 
 @dataclass
@@ -107,6 +118,8 @@ class ContentsWriter:
     contents: bytes
 
     def write_boot_file(self, path: Path) -> None:
+        if path.exists():
+            return
         with tempfile.NamedTemporaryFile(
             mode="wb",
             dir=path.parent,
@@ -590,8 +603,7 @@ def write_boot_files(boot_files: BootFileList) -> None:
     for boot_file in boot_files:
         boot_path = BOOT_MOUNT_POINT / boot_file.path
         try:
-            if not boot_path.exists():
-                boot_file.writer.write_boot_file(boot_path)
+            boot_file.writer.write_boot_file(boot_path)
         except OSError as e:
             # See https://github.com/NixOS/nixpkgs/issues/114552
             if e.errno == errno.EINVAL:
